+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 0.3
 * Add support for Bazaar.
 * Improve branch detection of Git.
 * Improve change detection of Git.
+* Fix root detection of Subversion.
 
 
 Version 0.2